Last adjustments to controller logic and error handling

+++ Carglass.TechnicalAssessment.Backend.Api/Controllers/ClientsController.cs
@@ -1,5 +1,6 @@
 using Carglass.TechnicalAssessment.Backend.BL;
 using Carglass.TechnicalAssessment.Backend.Dtos;
+using Carglass.TechnicalAssessment.Backend.Entities;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Carglass.TechnicalAssessment.Backend.Api.Controllers;
@@ -18,14 +19,23 @@ public class ClientsController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
-        return Ok(await _clientAppService.GetAll());
+        var clients = await _clientAppService.GetAll();
+
+        return Ok(clients);
     }
 
     [HttpGet]
     [Route("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
-        return Ok(await _clientAppService.GetById(id));
+        var client = await _clientAppService.GetById(id);
+
+        if (client is null)
+        {
+            return NotFound($"El cliente con el id: {id} no ha sido encontrado");
+        }
+
+        return Ok(client);
     }
 
     [HttpPost]
@@ -33,31 +43,52 @@ public class ClientsController : ControllerBase
     {
         await _clientAppService.Create(clientDto);
 
-        return Ok();
+        return CreatedAtAction(nameof(Create), clientDto);
     }
 
     [HttpPut]
     public async Task<IActionResult> Update([FromBody] ClientDto clientDto)
     {
+        var client = await _clientAppService.GetById(clientDto.Id);
+
+        if (client is null)
+        {
+            return BadRequest($"El cliente con el  id: {clientDto.Id} no existe");
+        }
+
         await _clientAppService.Update(clientDto);
 
-        return Ok();
-    }
-
-    [HttpDelete]
-    public async Task<IActionResult> Delete([FromBody] ClientDto clientDto)
-    {
-        await _clientAppService.Delete(clientDto);
-
-        return Ok();
+        return NoContent();
     }
 
     [HttpDelete]
     [Route("{id}")]
     public async Task<IActionResult> DeleteById(int id)
     {
+        var client = await _clientAppService.GetById(id);
+
+        if (client is null)
+        {
+            return BadRequest($"El cliente con el  id: {id} no existe");
+        }
+
         await _clientAppService.DeleteById(id);
 
-        return Ok();
+        return NoContent();
+    }
+
+    [HttpDelete]
+    public async Task<IActionResult> Delete([FromBody] ClientDto clientDto)
+    {
+        var client = await _clientAppService.GetById(clientDto.Id);
+
+        if (client is null)
+        {
+            return BadRequest("El cliente que desea borrar no existe");
+        }
+
+        await _clientAppService.Delete(clientDto);
+
+        return NoContent();
     }
 }
